chore: general code cleanups

+++ include/chesstillo/move_gen.hpp
@@ -1,6 +1,7 @@
 #ifndef MOVE_GEN_HPP
 #define MOVE_GEN_HPP
 
+#include <cstdint>
 #include <tuple>
 #include <vector>
 
@@ -39,7 +40,7 @@
    MOVE_NORTH_WEST(bitboard) | MOVE_SOUTH_EAST(bitboard) |                     \
    MOVE_SOUTH_WEST(bitboard))
 
-#define LOOP_INDEX static_cast<int>(_loop_i)
+#define LOOP_INDEX static_cast<uint8_t>(_loop_i)
 #define BITLOOP(X)                                                             \
   for (Bitboard _x_copy = X, _loop_i = BIT_INDEX(_x_copy); _x_copy;            \
        _x_copy ^= BITBOARD_FOR_SQUARE(LOOP_INDEX),                             \
@@ -50,21 +51,21 @@ std::tuple<Bitboard, Bitboard> PinMask(Position &position);
 
 std::vector<Move> GenerateMoves(Position &position);
 
-void AddMovesToList(std::vector<Move> &moves, int from, Bitboard targets,
+void AddMovesToList(std::vector<Move> &moves, uint8_t from, Bitboard targets,
                     Piece piece, Piece *mailbox, Bitboard enemy_bb);
 
-constexpr Bitboard RankMask(int square) { return kRank1 << (square & 56); }
+constexpr Bitboard RankMask(uint8_t square) { return kRank1 << (square & 56); }
 
-constexpr Bitboard FileMask(int square) { return kAFile << (square & 7); }
+constexpr Bitboard FileMask(uint8_t square) { return kAFile << (square & 7); }
 
-constexpr Bitboard DiagonalMask(int square) {
+constexpr Bitboard DiagonalMask(uint8_t square) {
   int diagonal = (square & 7) - (square >> 3);
 
   return diagonal >= 0 ? kA1H8Diagonal >> diagonal * 8
                        : kA1H8Diagonal << -diagonal * 8;
 }
 
-constexpr Bitboard AntiDiagonalMask(int square) {
+constexpr Bitboard AntiDiagonalMask(uint8_t square) {
   int diagonal = 7 - (square & 7) - (square >> 3);
 
   return diagonal >= 0 ? kH1A8Diagonal >> diagonal * 8
@@ -92,13 +93,13 @@ constexpr Bitboard SquaresInBetween(unsigned int from, unsigned int to) {
 }
 
 constexpr Bitboard BishopXRayAttacks(Bitboard attacks, Bitboard occupied_sqs,
-                                     Bitboard blockers, unsigned int square) {
+                                     Bitboard blockers, uint8_t square) {
   return attacks ^
          kSlidingAttacks.Bishop(occupied_sqs ^ (blockers & attacks), square);
 }
 
 constexpr Bitboard RookXRayAttacks(Bitboard attacks, Bitboard occupied_sqs,
-                                   Bitboard blockers, unsigned int square) {
+                                   Bitboard blockers, uint8_t square) {
   return attacks ^
          kSlidingAttacks.Rook(occupied_sqs ^ (blockers & attacks), square);
 }
