+++ src/position_test.cpp
@@ -1,9 +1,11 @@
 #include <array>
+#include <cstdint>
 
 #include <chesstillo/board.hpp>
 #include <chesstillo/fen.hpp>
 #include <chesstillo/position.hpp>
 #include <chesstillo/types.hpp>
+#include <chesstillo/utility.hpp>
 #include <gtest/gtest.h>
 
 class PositionTestSuite : public ::testing::Test {
@@ -27,7 +29,7 @@ TEST_F(PositionTestSuite, ApplyPawnMove) {
 }
 
 TEST_F(PositionTestSuite, ApplyPawnCaptureMove) {
-  std::array<std::array<int, 2>, 6> moves = {
+  std::array<std::array<uint8_t, 2>, 6> moves = {
       {{e2, e4}, {d7, d5}, {e4, d5}, {c7, c6}, {d2, d4}, {c6, d5}}};
 
   for (auto [from, to] : moves) {
@@ -51,7 +53,7 @@ TEST_F(PositionTestSuite, MoveKnight) {
 TEST_F(PositionTestSuite, MoveBishop) {
   ApplyFen(position, "8/8/8/3b4/3B4/8/8/8 w - - 0 1");
 
-  std::array<std::array<int, 2>, 2> moves = {{{d4, b6}, {d5, f7}}};
+  std::array<std::array<uint8_t, 2>, 2> moves = {{{d4, b6}, {d5, f7}}};
 
   for (auto [from, to] : moves) {
     Move move = DeduceMove(position, from, to);
@@ -65,7 +67,7 @@ TEST_F(PositionTestSuite, MoveBishop) {
 TEST_F(PositionTestSuite, MoveRook) {
   ApplyFen(position, "8/2r2n2/8/8/3R4/8/8/5R2 w - - 0 1");
 
-  std::array<std::array<int, 2>, 2> moves = {{{f1, f7}, {c7, f7}}};
+  std::array<std::array<uint8_t, 2>, 2> moves = {{{f1, f7}, {c7, f7}}};
 
   for (auto [from, to] : moves) {
     Move move = DeduceMove(position, from, to);
@@ -79,7 +81,7 @@ TEST_F(PositionTestSuite, MoveRook) {
 TEST_F(PositionTestSuite, MoveQueen) {
   ApplyFen(position, "8/8/4r3/1n1N2q1/2Q5/8/8/8 w - - 0 1");
 
-  std::array<std::array<int, 2>, 9> moves = {{
+  std::array<std::array<uint8_t, 2>, 9> moves = {{
       {c4, b5},
       {e6, e5},
       {b5, f1},
